+++ ignite/services/chain/serve.go
@@ -67,6 +67,7 @@ type serveOptions struct {
 	forceReset      bool
 	resetOnce       bool
 	skipProto       bool
+	skipBuild       bool
 	quitOnFail      bool
 	generateClients bool
 	buildTags       []string
@@ -114,6 +115,14 @@ func GenerateClients() ServeOption {
 func ServeSkipProto() ServeOption {
 	return func(c *serveOptions) {
 		c.skipProto = true
+	}
+}
+
+// ServeSkipBuild allows to serve the app without rebuilding it.
+// It looks up the binary in the PATH.
+func ServeSkipBuild() ServeOption {
+	return func(c *serveOptions) {
+		c.skipBuild = true
 	}
 }
 
@@ -188,6 +197,7 @@ func (c *Chain) Serve(ctx context.Context, cacheStorage cache.Storage, options .
 					serveOptions.buildTags,
 					shouldReset,
 					serveOptions.skipProto,
+					serveOptions.skipBuild,
 					serveOptions.generateClients,
 				)
 				serveOptions.resetOnce = false
@@ -317,7 +327,7 @@ func (c *Chain) serve(
 	ctx context.Context,
 	cacheStorage cache.Storage,
 	buildTags []string,
-	forceReset, skipProto, generateClients bool,
+	forceReset, skipProto, skipBuild, generateClients bool,
 ) error {
 	conf, err := c.Config()
 	if err != nil {
@@ -397,7 +407,13 @@ func (c *Chain) serve(
 	}
 
 	// build phase
-	if !isInit || appModified {
+	// if the app is not initialized or the source/binary has been modified
+	// and if the --skip-build flag is not set
+	if skipBuild {
+		c.ev.SendInfo("Skip building activated. Binary won't be rebuilt, nor refresh on changes")
+	}
+
+	if (!isInit || appModified) && !skipBuild {
 		// build the blockchain app
 		if err := c.build(ctx, cacheStorage, buildTags, "", skipProto, generateClients, true); err != nil {
 			return err