+++ changelog.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Features
+
+- [#4457](https://github.com/ignite/cli/pull/4457) Add `skip-build` flag to `chain serve` command to avoid (re)building the chain
+
 ## [`v28.6.1`](https://github.com/ignite/cli/releases/tag/v28.6.1)
 
 ### Changes